refactoring and testing validator
