+++ nexus_constructor/validators.py
@@ -12,6 +12,9 @@ from nexusutils.readwriteoff import parse_off_file
 from stl import mesh
 
 
+HDF_FILE_EXTENSIONS = ("nxs", "hdf", "hdf5")
+
+
 class NullableIntValidator(QIntValidator):
     """
     A validator that accepts integers as well as empty input.
@@ -402,5 +405,26 @@ class HDFLocationExistsValidator(QValidator):
     def _emit_and_return(self, valid: bool) -> QValidator.State:
         self.is_valid.emit(valid)
         return QValidator.Acceptable if valid else QValidator.Intermediate
+
+    is_valid = Signal(bool)
+
+
+class CommandDialogOKButtonValidator(QValidator):
+    """
+    Validator to ensure item names are unique within a model that has a 'name' property
+
+    The validationFailed signal is emitted if an entered name is not unique.
+    """
+
+    def __init__(self):
+        super().__init__()
+
+    def validate(self, input: str, pos: int) -> QValidator.State:
+        if not input or not input.endswith(HDF_FILE_EXTENSIONS):
+            self.is_valid.emit(False)
+            return QValidator.Intermediate
+
+        self.is_valid.emit(True)
+        return QValidator.Acceptable
 
     is_valid = Signal(bool)
